Former-commit-id: 056592daa1ff447145dd07c703bdae9cf366b863 [formerly 9c54ea5b5603b0274cb44b610025c8aae5014a87] [formerly 5894102a1bf7087d54c28fdb44bfb39655a0dd17] [formerly 056592daa1ff447145dd07c703bdae9cf366b863 [formerly 9c54ea5b5603b0274cb44b610025c8aae5014a87] [formerly 5894102a1bf7087d54c28fdb44bfb39655a0dd17] [formerly ec6502df640010eb1543078168a10310b8e6fb76 [formerly 5894102a1bf7087d54c28fdb44bfb39655a0dd17 [formerly 972e3d5117a3df147eaa83429f3649b1c1fb7a3f]]]]
Former-commit-id: ec6502df640010eb1543078168a10310b8e6fb76
Former-commit-id: 916f0bbe49e799139df916d4a85488eb021c753f [formerly 71fbcb15376b2e15a93eba05f568fc5b85e27e96] [formerly b54f72b4007bdc1417b21cfad790d8cdfa05ef2a [formerly e1a4a48258feeca54d59985c01d6a0118d7f0eb6]]
Former-commit-id: 7ce6d020324fb4838133cfb05fe3af33466efd1b [formerly 73f7caea8b6be3b5ddd1fd166bba6d1efd910290]
Former-commit-id: ab39d4d2a564cdadf792520da7d43c46cae04036

+++ edexOsgi/com.raytheon.edex.plugin.gfe/src/com/raytheon/edex/plugin/gfe/server/database/IFPGridDatabase.java
@@ -115,6 +115,7 @@ import com.raytheon.uf.edex.database.DataAccessLayerException;
  * 10/31/2013   #2508      randerso    Change to use DiscreteGridSlice.getKeys()
  * 12/10/13     #2611      randerso    Change saveGridData to set update time when saving grids
  * 05/29/2014   #3071      randerso    Fix NPE in getCachedParmID
+ * 09/21/2014   #3648      randerso    Changed deleteDatabase to handle database already being deleted by other JVM
  * 
  * </pre>
  * 
@@ -2038,15 +2039,18 @@ public class IFPGridDatabase extends GridDatabase {
      *            the DatabaseID of the datbase to be deleted
      */
     public static void deleteDatabase(DatabaseID id) {
+        boolean purged = false;
         try {
             GFEDao gfeDao = new GFEDao();
-            gfeDao.purgeGFEGrids(id);
+            purged = gfeDao.purgeGFEGrids(id);
         } catch (Exception e) {
             statusHandler.handle(Priority.PROBLEM,
                     "Unable to delete model database: " + id, e);
         }
 
-        deleteModelHDF5(id);
+        if (purged) {
+            deleteModelHDF5(id);
+        }
     }
 
     /**
